URL bug fix Attempt #1
Because the index file in the root directory have only an iframe tag and all functioning pages are inside the src folder, when binod.website is being opened no clicks are changing the url of the page, which kanna looks cool, but may deliver a bad user experience. So I am adding parent target to every links.